% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtp.R
\name{mmtp}
\alias{mmtp}
\title{Causal Mediation Using Modified Treatment Policies}
\usage{
mmtp(data, vars, d_prime, d_star, id = NULL, control = .lcmmtp_control())
}
\arguments{
\item{data}{A data frame containing all necessary variables}

\item{vars}{An `mmtp_variables` object mapping observed variables to
the assumed variable structure.}

\item{control}{}
}
\value{
An object of class `mmtp`
}
\description{
Causal Mediation Using Modified Treatment Policies
}
\examples{
vars <- mmtp_variables$new(
   W = list(trt = c("L_1"), outcome = c("L_1")),
   A = "A_1",
   Z = "Z_1",
   M = "M_1",
   Y = "Y",
   cens = c("c2")
)

d_ap <- function(data, trt) rep(1, length(data[[trt]]))
d_as <- function(data, trt) rep(0, length(data[[trt]]))

mmtp(lcmmtp::lcmmtp_foo, vars, d_ap, d_as, id = NULL, .lcmmtp_control(folds = 2))
}
