% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtp_variables.R
\name{mmtp_variables}
\alias{mmtp_variables}
\title{R6 class for an lcmmtp_variables}
\description{
R6 class for an lcmmtp_variables

R6 class for an lcmmtp_variables
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mmtp_variables-new}{\code{mmtp_variables$new()}}
\item \href{#method-mmtp_variables-history}{\code{mmtp_variables$history()}}
\item \href{#method-mmtp_variables-all_vars}{\code{mmtp_variables$all_vars()}}
\item \href{#method-mmtp_variables-clone}{\code{mmtp_variables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mmtp_variables-new"></a>}}
\if{latex}{\out{\hypertarget{method-mmtp_variables-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mmtp_variables$new(W, A, Z, M, Y, cens)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mmtp_variables-history"></a>}}
\if{latex}{\out{\hypertarget{method-mmtp_variables-history}{}}}
\subsection{Method \code{history()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mmtp_variables$history(var = c("A", "Z", "M", "Y"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mmtp_variables-all_vars"></a>}}
\if{latex}{\out{\hypertarget{method-mmtp_variables-all_vars}{}}}
\subsection{Method \code{all_vars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mmtp_variables$all_vars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mmtp_variables-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mmtp_variables-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mmtp_variables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
